+++ domain/src/main/java/com/mordansoft/homebank/domain/usecase/purchase/GetMainPurchasesUc.kt
@@ -8,7 +8,7 @@ class GetMainPurchasesUc(val purchaseRepo: PurchaseRepo) {
     suspend fun execute(): ArrayList<Purchase> {
         purchaseRepo.insertTestPurchase()
         val parentId : Long = -8
-        val periodId : Long = -8
+        val periodId : Int = -8
         return purchaseRepo.getMainPurchases(parentId = parentId,
                                              periodId = periodId)
     }
